refactor: use a mediatr command to seed the tier
