+++ ConsumerApi/QuotasDbContextSeeder.cs
@@ -1,10 +1,10 @@
 ﻿using Backbone.BuildingBlocks.API.Extensions;
-using Backbone.BuildingBlocks.Application.Abstractions.Exceptions;
 using Backbone.Modules.Devices.Infrastructure.Persistence.Database;
-using Backbone.Modules.Quotas.Application.Infrastructure.Persistence.Repository;
+using Backbone.Modules.Quotas.Application.Tiers.Commands.SeedUpForDeletionTier;
 using Backbone.Modules.Quotas.Domain.Aggregates.Identities;
 using Backbone.Modules.Quotas.Domain.Aggregates.Tiers;
 using Backbone.Modules.Quotas.Infrastructure.Persistence.Database;
+using MediatR;
 using Microsoft.EntityFrameworkCore;
 
 namespace Backbone.ConsumerApi;
@@ -12,14 +12,12 @@ namespace Backbone.ConsumerApi;
 public class QuotasDbContextSeeder : IDbSeeder<QuotasDbContext>
 {
     private readonly DevicesDbContext _devicesDbContext;
-    private readonly IMetricsRepository _metricsRepository;
-    private readonly ITiersRepository _tiersRepository;
+    private readonly IMediator _mediator;
 
-    public QuotasDbContextSeeder(DevicesDbContext devicesDbContext, IMetricsRepository metricsRepository, ITiersRepository tiersRepository)
+    public QuotasDbContextSeeder(DevicesDbContext devicesDbContext, IMediator mediator)
     {
         _devicesDbContext = devicesDbContext;
-        _metricsRepository = metricsRepository;
-        _tiersRepository = tiersRepository;
+        _mediator = mediator;
     }
 
     public async Task SeedAsync(QuotasDbContext context)
@@ -31,25 +29,7 @@ public class QuotasDbContextSeeder : IDbSeeder<QuotasDbContext>
 
     private async Task EnsureUpForDeletionTierWithQuotas(QuotasDbContext context)
     {
-        Tier upForDeletionTier;
-
-        try
-        {
-            upForDeletionTier = await _tiersRepository.Find(TierId.UP_FOR_DELETION_DEFAULT_ID, CancellationToken.None, true);
-        }
-        catch (NotFoundException)
-        {
-            upForDeletionTier = new Tier(new TierId(TierId.UP_FOR_DELETION_DEFAULT_ID), Tier.UP_FOR_DELETION_TIER_NAME);
-            await _tiersRepository.Add(upForDeletionTier, CancellationToken.None);
-        }
-
-        var metrics = await _metricsRepository.FindAll(CancellationToken.None);
-        var missingMetrics = metrics.Where(metric => upForDeletionTier.Quotas.All(quota => quota.MetricKey.Value != metric.Key.Value));
-        foreach (var metric in missingMetrics)
-        {
-            upForDeletionTier.CreateQuotaForUpForDeletionTier(metric.Key, 0, QuotaPeriod.Day);
-        }
-        await _tiersRepository.Update(upForDeletionTier, CancellationToken.None);
+        await _mediator.Send(new SeedUpForDeletionTierCommand());
     }
 
     private async Task AddTierToIdentities(QuotasDbContext context)
