+++ Modules/Quotas/src/Quotas.Application/Tiers/Commands/SeedUpForDeletionTier/SeedUpForDeletionTierCommand.cs
@@ -1,0 +1,7 @@
+﻿using MediatR;
+
+namespace Backbone.Modules.Quotas.Application.Tiers.Commands.SeedUpForDeletionTier;
+
+public class SeedUpForDeletionTierCommand : IRequest
+{
+}
